% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guessSex.R
\name{guessSex}
\alias{guessSex}
\title{Guesses the sex of samples from a bsseq object}
\usage{
guessSex(bs)
}
\arguments{
\item{bs}{A bsseq object, must include CpGs on sex (X and Y) chromosomes.}
}
\description{
Guesses the sex of samples from a bsseq object
}
\examples{
guessSex(bs)
}
