% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictAge.R
\name{predictAge}
\alias{predictAge}
\title{Predicts the age  of samples from a bsseq object}
\usage{
predictAge(bs, adult.age = 20, fastImpute = FALSE, imp = 10)
}
\arguments{
\item{bs}{A bsseq object}

\item{adult.age}{Adult age for transformation. Default is 20}

\item{fastImpute}{Whether fast imptation should be performed. This is not advised as it leads to less accurate predictions. Default is FALSE}

\item{imp}{Number of imputations to be performed by mice() during imputation of missing methylation values. Default is 10}
}
\description{
Predicts the age  of samples from a bsseq object
}
\examples{
predictAge(bs, adult.age = 20, fastImpute = FALSE, imp = 10)
}
